authentication/pyproject.toml
[project]
name = "authentication"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.5",
    "fastapi>=0.116.0",
    "passlib[bcrypt]>=1.7.4",
    "pymysql>=1.1.1",
    "python-jose[crptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
